Made more animations use hue/saturation/brightness. Fixed a few that had typos. Fixed an issue with the bottom torus pixel not being set correctly, which fixed an animation that was causing the board to freeze (fourSquare). Fixed frequency offset to clamp correctly. Clamped the animation speed to a reasonable range.

+++ Equalizer.h
@@ -18,8 +18,8 @@ public:
     void readAudioFrequencies();
     uint8_t getBand(uint8_t band);
     uint8_t getBand(uint8_t band, uint16_t sensitivityThreshold);
-    uint8_t getFrequencyOffset();
-    void setFrequencyOffset(uint8_t newFrequencyOffset);
+    int getFrequencyOffset();
+    void setFrequencyOffset(int newFrequencyOffset);
     uint16_t getSensitivity();
     void setSensitivity(uint16_t newSensitivity);
     void resetFrequencyOffset();
